refactor: transaction.goを4つのファイルに分割
RFC32161の17章で定義されている4種類のトランザクション（INVITE Client Transaction、Non-INVITE Client Transaction、INVITE Server Transaction、Non-INVITE Server Transaction）について、これまで`transaction.go`にまとめられていた処理を、それぞれ以下のファイルに分割しました。

- `transaction_invite_server.go`
- `transaction_noninvite_server.go`
- `transaction_invite_client.go`
- `transaction_noninvite_client.go`

`transaction.go`には、トランザクション間で共通のインターフェースや定数、TransactionManagerのみが残るようにリファクタリングしています。
